Add the classifier as proposed by Schuld, Fingerhuth, Petruccione
I'm repeating the work described in the article to better understand the techniques and to get a solid base to further expand.
The procedure with the QRAM previously implemented in this repository didn't yield clear results
when applied to a bigger number of qubits, and I'd like to pursue different roads to spot the shortcomings.