Automate the vector encoding
Encoding both input and training vectors is now more automated.
The class qubits are inferred from the name of the class, and you
can write the index qubit just changing a number (only 3 m qubits for now).
It is also easier to pick a random input vector from each class and analyze it.